requirements.txt
mysql-connector-python
Flask
Flask-Cors
python-dotenv